+++ main.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 )
 
-const version = "1.1.0"
+const version = "1.2.0"
 
 type options struct {
 	action      string